requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.13.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
